+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @proxima-io/cli
 
+## 0.3.1
+
+### Patch Changes
+
+- `proxima deploy --auto-scaffold` for smart collection auto-scaffold during website deploy.
+- Updated dependencies
+  - @proxima-io/templateizer@0.2.2
+
 ## 0.3.0
 
 ### Minor Changes